fix: move dfs strategy in another module
