+++ libs/core/include/cynthia/sdd_to_formula.hpp
@@ -21,8 +21,7 @@
 namespace cynthia {
 namespace core {
 
-logic::ltlf_ptr sdd_to_formula(SddNode* sdd_node,
-                               ForwardSynthesis::Context& context_);
+logic::ltlf_ptr sdd_to_formula(SddNode* sdd_node, Context& context_);
 
 } // namespace core
 } // namespace cynthia
