+++ libs/core/tests/unit/test_core/test_complex_formulas.cpp
@@ -18,6 +18,7 @@
 #include <catch.hpp>
 #include <cynthia/core.hpp>
 #include <cynthia/parser/driver.hpp>
+#include <cynthia/strategies/dfs.hpp>
 #include <sstream>
 
 namespace cynthia {
@@ -34,7 +35,7 @@ TEST_CASE("forward synthesis of random formula 1") {
   auto not_end = temp->ctx().make_not_end();
   auto formula = temp->ctx().make_and({temp, not_end});
   auto partition = InputOutputPartition({"p5"}, {"p0", "p1", "p3", "p4"});
-  bool result = is_realizable<ForwardSynthesis>(formula, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
   REQUIRE(result);
 }
 
@@ -47,7 +48,7 @@ TEST_CASE("forward synthesis of random formula 2") {
   auto not_end = temp->ctx().make_not_end();
   auto formula = temp->ctx().make_and({temp, not_end});
   auto partition = InputOutputPartition({"p1", "p0", "p4"}, {"p3"});
-  bool result = is_realizable<ForwardSynthesis>(formula, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
   REQUIRE(result);
 }
 TEST_CASE("forward synthesis of random formula 3") {
@@ -60,7 +61,7 @@ TEST_CASE("forward synthesis of random formula 3") {
   auto not_end = temp->ctx().make_not_end();
   auto formula = temp->ctx().make_and({temp, not_end});
   auto partition = InputOutputPartition({"p5", "p3"}, {"p1", "p0", "p4"});
-  bool result = is_realizable<ForwardSynthesis>(formula, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
   REQUIRE(result);
 }
 TEST_CASE("forward synthesis of random formula 4") {
@@ -72,7 +73,7 @@ TEST_CASE("forward synthesis of random formula 4") {
   auto not_end = temp->ctx().make_not_end();
   auto formula = temp->ctx().make_and({temp, not_end});
   auto partition = InputOutputPartition({"p2", "p4"}, {"p1", "p0"});
-  bool result = is_realizable<ForwardSynthesis>(formula, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
   REQUIRE(result);
 }
 TEST_CASE("forward synthesis of random formula 5") {
@@ -84,7 +85,7 @@ TEST_CASE("forward synthesis of random formula 5") {
   auto not_end = temp->ctx().make_not_end();
   auto formula = temp->ctx().make_and({temp, not_end});
   auto partition = InputOutputPartition({"p0"}, {"dummy"});
-  bool result = is_realizable<ForwardSynthesis>(formula, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
   REQUIRE(!result);
 }
