+++ libs/core/tests/unit/test_core/test_core_basic_formulas.cpp
@@ -17,6 +17,7 @@
 
 #include <catch.hpp>
 #include <cynthia/core.hpp>
+#include <cynthia/strategies/dfs.hpp>
 
 namespace cynthia {
 namespace core {
@@ -26,7 +27,7 @@ TEST_CASE("forward synthesis of 'tt'") {
   logic::Context context;
   auto partition = InputOutputPartition({"a"}, {"b"});
   auto tt = context.make_tt();
-  bool result = is_realizable<ForwardSynthesis>(tt, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(tt, partition);
   REQUIRE(result);
 }
 
@@ -34,7 +35,7 @@ TEST_CASE("forward synthesis of 'ff'") {
   logic::Context context;
   auto partition = InputOutputPartition({"a"}, {"b"});
   auto ff = context.make_ff();
-  bool result = is_realizable<ForwardSynthesis>(ff, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(ff, partition);
   REQUIRE(!result);
 }
 
@@ -42,7 +43,7 @@ TEST_CASE("forward synthesis of 'true'") {
   logic::Context context;
   auto partition = InputOutputPartition({"a"}, {"b"});
   auto true_ = context.make_prop_true();
-  bool result = is_realizable<ForwardSynthesis>(true_, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(true_, partition);
   REQUIRE(result);
 }
 
@@ -50,7 +51,7 @@ TEST_CASE("forward synthesis of 'false'") {
   logic::Context context;
   auto partition = InputOutputPartition({"a"}, {"b"});
   auto false_ = context.make_prop_false();
-  bool result = is_realizable<ForwardSynthesis>(false_, partition);
+  bool result = is_realizable<ForwardDfsSynthesis>(false_, partition);
   REQUIRE(!result);
 }
 TEST_CASE("forward synthesis of atom") {
@@ -59,12 +60,12 @@ TEST_CASE("forward synthesis of atom") {
 
   SECTION("atom controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a, partition);
     REQUIRE(result);
   }
   SECTION("atom not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a, partition);
     REQUIRE(!result);
   }
 }
@@ -75,12 +76,12 @@ TEST_CASE("forward synthesis of not atom") {
 
   SECTION("atom controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(not_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(not_a, partition);
     REQUIRE(result);
   }
   SECTION("atom not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(not_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(not_a, partition);
     REQUIRE(!result);
   }
 }
@@ -93,22 +94,22 @@ TEST_CASE("forward synthesis of 'a and b'") {
 
   SECTION("left not-controllable, right not-controllable") {
     auto partition = InputOutputPartition({"a", "b"}, {"c"});
-    bool result = is_realizable<ForwardSynthesis>(a_and_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_and_b, partition);
     REQUIRE(!result);
   }
   SECTION("left not-controllable, right controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(a_and_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_and_b, partition);
     REQUIRE(!result);
   }
   SECTION("left controllable, right not-controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(a_and_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_and_b, partition);
     REQUIRE(!result);
   }
   SECTION("left controllable, right controllable") {
     auto partition = InputOutputPartition({"c"}, {"a", "b"});
-    bool result = is_realizable<ForwardSynthesis>(a_and_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_and_b, partition);
     REQUIRE(result);
   }
 }
@@ -120,22 +121,22 @@ TEST_CASE("forward synthesis of 'a or b'") {
   auto a_or_b = context.make_or({a, b});
   SECTION("left not-controllable, right not-controllable") {
     auto partition = InputOutputPartition({"a", "b"}, {"c"});
-    bool result = is_realizable<ForwardSynthesis>(a_or_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_or_b, partition);
     REQUIRE(!result);
   }
   SECTION("left not-controllable, right controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(a_or_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_or_b, partition);
     REQUIRE(result);
   }
   SECTION("left controllable, right not-controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(a_or_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_or_b, partition);
     REQUIRE(result);
   }
   SECTION("left controllable, right controllable") {
     auto partition = InputOutputPartition({"c"}, {"a", "b"});
-    bool result = is_realizable<ForwardSynthesis>(a_or_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_or_b, partition);
     REQUIRE(result);
   }
 }
@@ -147,12 +148,12 @@ TEST_CASE("forward synthesis of 'X[!]a' (controllable)") {
 
   SECTION("body controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(next_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(next_a, partition);
     REQUIRE(result);
   }
   SECTION("body not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(next_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(next_a, partition);
     REQUIRE(!result);
   }
 }
@@ -164,12 +165,12 @@ TEST_CASE("forward synthesis of 'X a'") {
 
   SECTION("body controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(weak_next_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(weak_next_a, partition);
     REQUIRE(result);
   }
   SECTION("body not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(weak_next_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(weak_next_a, partition);
     REQUIRE(result);
   }
 }
@@ -182,12 +183,12 @@ TEST_CASE("forward synthesis of 'X a', non-empty trace") {
 
   SECTION("body controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("body not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
 }
@@ -199,22 +200,22 @@ TEST_CASE("forward synthesis of 'a U b'") {
 
   SECTION("head non-controllable, tail non-controllable") {
     auto partition = InputOutputPartition({"a", "b"}, {"c"});
-    bool result = is_realizable<ForwardSynthesis>(a_until_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_until_b, partition);
     REQUIRE(!result);
   }
   SECTION("head non-controllable, tail controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(a_until_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_until_b, partition);
     REQUIRE(result);
   }
   SECTION("head controllable, tail non-controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(a_until_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_until_b, partition);
     REQUIRE(!result);
   }
   SECTION("head controllable, tail controllable") {
     auto partition = InputOutputPartition({"c"}, {"a", "b"});
-    bool result = is_realizable<ForwardSynthesis>(a_until_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_until_b, partition);
     REQUIRE(result);
   }
 }
@@ -228,22 +229,22 @@ TEST_CASE("forward synthesis of 'a R b'") {
 
   SECTION("head non-controllable, tail non-controllable") {
     auto partition = InputOutputPartition({"a", "b"}, {"c"});
-    bool result = is_realizable<ForwardSynthesis>(a_release_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_release_b, partition);
     REQUIRE(result);
   }
   SECTION("head non-controllable, tail controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(a_release_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_release_b, partition);
     REQUIRE(result);
   }
   SECTION("head controllable, tail non-controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(a_release_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_release_b, partition);
     REQUIRE(result);
   }
   SECTION("head controllable, tail controllable") {
     auto partition = InputOutputPartition({"c"}, {"a", "b"});
-    bool result = is_realizable<ForwardSynthesis>(a_release_b, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(a_release_b, partition);
     REQUIRE(result);
   }
 }
@@ -255,12 +256,12 @@ TEST_CASE("forward synthesis of 'F a'") {
 
   SECTION("tail not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(eventually_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(eventually_a, partition);
     REQUIRE(!result);
   }
   SECTION("tail controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(eventually_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(eventually_a, partition);
     REQUIRE(result);
   }
 }
@@ -273,12 +274,12 @@ TEST_CASE("forward synthesis of 'G a'") {
 
   SECTION("tail not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(always_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(always_a, partition);
     REQUIRE(result);
   }
   SECTION("tail controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(always_a, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(always_a, partition);
     REQUIRE(result);
   }
 }
@@ -291,12 +292,12 @@ TEST_CASE("forward synthesis of 'G a', non-empty trace") {
 
   SECTION("tail not-controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(!result);
   }
   SECTION("tail controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
 }
@@ -312,12 +313,12 @@ TEST_CASE("forward synthesis of 'G(a | !b)', non-empty trace") {
 
   SECTION("b controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("a controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
 }
@@ -331,12 +332,12 @@ TEST_CASE("forward synthesis of 'X(F(p0))'") {
   auto formula = context.make_and({next, not_end});
   SECTION("p0 uncontrollable") {
     auto partition = InputOutputPartition({"p0"}, {"p1"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("p0 controllable") {
     auto partition = InputOutputPartition({"p1"}, {"p0"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
 }
@@ -355,12 +356,12 @@ TEST_CASE("forward synthesis of '((a) & (X(~(a)))) | ((~(a)) & (X(a)))'") {
 
   SECTION("a uncontrollable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("a controllable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
 }
@@ -376,22 +377,22 @@ TEST_CASE("forward synthesis of 'p0 R (F(p1))'") {
 
   SECTION("p0 controllable, p1 controllable") {
     auto partition = InputOutputPartition({"dummy"}, {"p0", "p1"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("p0 uncontrollable, p1 controllable") {
     auto partition = InputOutputPartition({"p0"}, {"p1"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("p0 controllable, p1 uncontrollable") {
     auto partition = InputOutputPartition({"p1"}, {"p0"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(!result);
   }
   SECTION("p0 uncontrollable, p1 uncontrollable") {
     auto partition = InputOutputPartition({"p0", "p1"}, {"dummy"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(!result);
   }
 }
@@ -409,22 +410,22 @@ TEST_CASE("forward synthesis of '(X(F(~b))) U (G(a))'") {
 
   SECTION("a controllable, b controllable") {
     auto partition = InputOutputPartition({"dummy"}, {"a", "b"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("a uncontrollable, b controllable") {
     auto partition = InputOutputPartition({"a"}, {"b"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(!result);
   }
   SECTION("a controllable, b uncontrollable") {
     auto partition = InputOutputPartition({"b"}, {"a"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(result);
   }
   SECTION("a uncontrollable, b uncontrollable") {
     auto partition = InputOutputPartition({"a", "b"}, {"dummy"});
-    bool result = is_realizable<ForwardSynthesis>(formula, partition);
+    bool result = is_realizable<ForwardDfsSynthesis>(formula, partition);
     REQUIRE(!result);
   }
 }
